[processing] keep only one zonal statistics algorithm
